+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - TabbedContent will automatically make tabs active when a widget in a pane is focused https://github.com/Textualize/textual/issues/4593
 
+### Added
+
+- Added hatch style https://github.com/Textualize/textual/pull/4603
+
 ## [0.64.0] - 2024-06-03
 
 ### Fixed